requirements.txt
Flask
lxml
gunicorn
celery
redis
python-dotenv
